server/requirements.txt
flask~=2.3.3
requests~=2.31.0
python-dotenv~=1.0.1
gunicorn ~=22.0.0
#Flask-SQLAlchemy==3.1.1
#psycopg-binary==3.1.18